match: carve GetStrTalkLen + EventA9_WmUnitMovePaths region-diff to C (Event team Worker D)
Full make compare OK. Last net-new matches from the team worktrees:
- sub_8008A40 GetStrTalkLen (968B): (s8) param-widen forces JP prologue sign-extend
  (lsls#24;asrs#24); scene.h proto s8->int (byte-neutral, all callers cold-rebuilt OK).
- sub_800CA34 EventA9_WmUnitMovePaths: int-local-widen.
Real JP addrs recovered by disasm (nofuncmap addr 0x08008B44 was stale; real 0x08008A40).

+++ include/scene.h
@@ -209,7 +209,7 @@ void PrintStringToTexts(struct Text** texts, const char* str, u16 * tm, int unk)
 void TalkPutSpriteText_OnIdle(struct Proc*);
 void ClearPrimaryHBlank(void);
 void TalkPutSpriteText_OnEnd(void);
-int GetStrTalkLen(const char*, s8);
+int GetStrTalkLen(const char*, int);
 bool GetZero(void);
 void TalkDebugNop(void);
 void TalkBgSync(int bg);
